Extract ClassMethod#stub_method_definition
And override the implementation in AnyInstanceMethod. I want to make the
two implementations of `#define_new_method` identical so that I can
remove the duplication.

Note that `stub_method_definition` is returning the method definition,
the filename and the line number so that exception backtraces continue to
report the correct line that raised the error. I've updated two tests
accordingly.