Encapsulate access to @original_method
TODO: Explain this commit. I initially tried lazily returning
`original_method` but that meant that it could potentially return the
stubbed implementation depending on when it was called.

Note the use of `original_method_in_scope`. I can't simply use
`original_method` within `define_method` as that's not available at the time
the block passed to `define_method` is evaluated.