Add tests for reporting filename and line number
We use `class_eval` in both `ClassMethod#define_new_method` and
`AnyInstanceMethod#define_new_method`. In both cases, we pass `__FILE__` and `__LINE__` to
`class_eval` so that the backtrace contains the filename and line number
of the line that raised the exception.

This commit adds a couple of tests to ensure we don't accidentally break
this behaviour. I'm not super happy with the tests (at the very least
they're brittle) but I want them in place before making some planned
changed to both implementations of `#define_new_method`.